Split convert.go into focused files
convert.go was a 1,005-line grab bag covering every resource kind. This
moves the code into smaller files organized by responsibility, with no
behavior change (all existing snapshot tests pass):

- annotations.go: exported annotation/label key constants
- convert.go: Transformer, Convert orchestration, validateService,
  getServiceName, mergeMaps, escapeEnvs (~200 lines)
- workloads.go: Pod / Deployment / DaemonSet / addContainersToSpec /
  deployment + daemonset update strategies
- pod_spec.go: createPodSpec + helpers (security context, topology
  spread, restart policy, match labels, extra hosts)
- containers.go: createContainer + per-container converters (env,
  resources, ports, pull policy) + removeDuplicateVolumeMounts
- services.go: Service / Ingress / ServiceAccount + service-port and
  protocol conversion
- probes.go: getProbes + getFirstPort + splitCommand

Also extracts a getServiceName helper that replaces four copies of the
"service group annotation else service name" pattern in createService,
createDeployment, createDaemonSet, and getMatchLabels. Method receivers
that never used the Transformer (convertPorts, convertServicePorts,
convertExtraHosts, getResourceRequirements, getImagePullPolicy) are now
plain package-level functions for consistency.
